Better behavior when the user sets .value on the knob that is being rotated.
Let's suppose the "user" (i.e. the developer who is using x-knob
component) writes an event handler for `input` event. Inside this
handler, the code sets a new `value` for the knob (e.g. `ev.target.value
+= 1`).

This commit tries to improve the handling of this case, but may cause
drifting. See the example included in this commit.